just put it in the build

+++ build/build.sh
@@ -1,5 +1,18 @@
 #!/bin/bash
 
+echo "Checking if npm is installed"
+## checking nvm version
+if [ -e ~/.nvm/nvm.sh ];
+then
+   source ~/.nvm/nvm.sh
+fi
+npm --version
+if [ $? -ne 0 ];
+then
+  echo "npm is not installed.  Try https://github.com/creationix/nvm"
+  exit 2
+fi
+
 echo "Building hapi demo"
 echo "Cleanup"
 rm -f *.tar.gz
